web/requirements.txt
django==4.2.2
django-bootstrap-v5==1.0.11
python-dotenv==1.0.0
opencv-python-headless==4.7.0.72
pillow==9.5.0
numpy==1.25.0
requests==2.28.2
gunicorn==20.1.0